Use gint64 instead of off_t for public APIs.
2008-05-23  Jeffrey Stedfast  <fejj@novell.com>

	* gmime/*.[c,h]: Use gint64 instead of off_t for public APIs.

	* mono/gmime-api.raw: s/off_t/gint64/g


svn path=/trunk/; revision=1247

+++ util/url-scanner.h
@@ -30,8 +30,8 @@ G_BEGIN_DECLS
 typedef struct {
 	const char *pattern;
 	const char *prefix;
-	off_t um_so;
-	off_t um_eo;
+	size_t um_so;
+	size_t um_eo;
 } urlmatch_t;
 
 typedef gboolean (*UrlScanFunc) (const char *in, const char *pos, const char *inend, urlmatch_t *match);
